Add methods where needed for Dart 2.0 core lib changes.
Right now, all of the implementations throw (except in trivial cases, like the
EmptyUnmodifiable* classes). They can be updated, typically to call `super` or
reference a wrapped object, once the new methods have been released as part of a
Dart 2.0 release, and this pubspec.yaml updates to guarantee that those
implementations exist.