Rewrite doc comments to refer to TypeError (#131)
This will be the type going forward. Keep a reference to `CastError` for
now, we can drop it after a few SDK releases.

Ignore usage of `CastError` in a test so that we can keep running on
the oldest supported stable SDK. We can switch this when the minimum
SDK is bumped.