Avoid public `late final` field
This had been set in the constructor because it needs to reference
a field on `this`. A new capability with late final files is to allow an
initializer expression referencing `this`. This approach also avoids
introducing a setter to the public API. Technically it's statically
breaking to remove the setter, but since it could never be called at
runtime this is safe to do.