Changes to Equality classes.
`MapEquality` and `SetEquality` now expects the provided
equalities to be compatible with the inherent equals
(and hash) of the map or set.
That allows using `Map.operator[]` and `Set.lookup`
to match entries/elements from one collection with
those of the other, instead of building an extra
hash-map on the side to count occurrences.

This helps `DeepCollectionEquality` which is likely
to traverse a larger collection structure (mainly maps)
and do repeated equals/hash code on sup-components.

For unordered `DeepCollectionEquality`, a cache of
computed values is added to avoid repeating the same
computation during the traversal of a map-structure.

Also adds `JsonEquality` which is tuned to JSON-like
map/list/basic-value structures, and which assumes
everything uses default `==` equality.