Make sure .sample()'s result is randomized in more cases
Sampling result was always not randomized when requested samples count
was equal or smaller than the input list length and was lots of times
not randomized when requested samples count was a little smaller
than input list length.

This fixes it simply by using core provided .shuffle() and emulates
sampling on top of it and provides tests that was failing on the
older code but now don't fail.