Make sure .sample()'s result is randomized in more cases (#302)
Samples which retain elements from early in the input would have those
elements in the same order as the input. As the sample size approaches
the size of the input, the ordering was more noticeably correlated to the
input order.

Use random indexes for the initial unconditionally added elements before
the later elements are selected into random indexes. This allows the
output order to be consistently uncorrelated with the input order, even
for samples of the entire input.