v2.0.0 breaking changes
Starts using Dart 3.0 class modifiers.
This is a breaking change, making some classes be `final` or `interface`
which prevents behavior that was previously possible.

Removes existing deprecated API, except for `whereNotNull`.

Deprecates API that is available in platform libraries (`IterableZip` class).

Changes behavior of `UnorderedIterableEquality`, `SetEquality` and `MapEquality`
to assume the compared collections' notions of equality agree with the
`Equality` object for elements/keys in the `UnorderedIterableEquality`/
`SetEquality`/`MapEquality` object.
This should improve performance of the equality comparisons, which currently
try to assume nothing, and therefore cannot use `.contains`/`.containsKey`/`[]`
on one collection with keys/values from the other collection